Fix minor issues; add error and compatibility checking

+++ bivariate/polynomial_test.go
@@ -13,8 +13,8 @@ var prg = rand.New(rand.NewSource(time.Now().UTC().UnixNano()))
 
 func TestAddAndSubDegs(t *testing.T) {
 	for i := 0; i < 1000; i++ {
-		a, b := uint(prg.Uint64()), uint(prg.Uint64())
-		c, d := uint(prg.Uint64()), uint(prg.Uint64())
+		a, b := uint(prg.Uint32()), uint(prg.Uint32())
+		c, d := uint(prg.Uint32()), uint(prg.Uint32())
 		if tmp, _ := addDegs([2]uint{a, b}, [2]uint{c, d}); tmp != [2]uint{a + c, b + d} {
 			t.Errorf("addDegs({%d,%d},{%d,%d})=%v (Expected {%d,%d})", a, b, c, d, tmp, a+c, b+d)
 		}
